test: move ProfilingPlugin Chrome E2E into the unit suite via puppeteer-core
Replace the dedicated latest-LTS CI job with a committed puppeteer-core
devDependency (no postinstall, so it installs cleanly on the legacy-Node
matrix) and run the browser check in the unit suite instead of the
integration matrix. It launches the runner's installed Chrome and self-skips
when none is available or under the Bun/Deno runtimes.
