+++ .github/workflows/test.yml
@@ -167,32 +167,6 @@ jobs:
         env:
           CODECOV_TOKEN: ${{ secrets.CODECOV_TOKEN }}
 
-  # ProfilingPlugin browser end-to-end check. puppeteer needs Node 18+ and pulls
-  # a Chromium download, so it is not a committed devDependency (that breaks the
-  # legacy-Node install matrix); install it on demand and run only here, on the
-  # latest LTS. The test self-skips everywhere puppeteer/Chrome is absent.
-  profiling-plugin-chrome:
-    needs: basic
-    runs-on: ubuntu-latest
-    steps:
-      - uses: actions/checkout@df4cb1c069e1874edd31b4311f1884172cec0e10 # v6.0.3
-
-      - name: Use Node.js
-        uses: actions/setup-node@48b55a011bda9f5d6aeb4c2d9c7362e8dae4041e # v6.4.0
-        with:
-          node-version: lts/*
-          cache: yarn
-
-      - run: yarn --frozen-lockfile
-
-      - run: yarn add --dev puppeteer@^23 --ignore-engines
-
-      - run: yarn link --frozen-lockfile || true
-
-      - run: yarn link webpack --frozen-lockfile
-
-      - run: yarn test:base test/ProfilingPlugin.longtest.js --ci
-
   test262:
     needs: basic
     strategy:
